requirements.txt
python-telegram-bot==21.6
huggingface_hub>=0.20.0
google-genai>=0.3.0
python-dotenv
pymupdf>=1.23.0
fpdf2>=2.7.0
